Corrected calls to SQLAlchemy (from one() to first()) and added item_delete

+++ pryce/database/dal/item.py
@@ -18,27 +18,25 @@ class DALItem:
 
     def get_item(self, code):
         item = None
-        try:
-            item = Item.query.filter_by(code=code).one()
-        except MultipleResultsFound as mrf:
-            raise MultipleResultsFound
+        item = Item.query.filter_by(code=code).first()
         return item
 
     def update_item(self, item_dict):
         item = None
         code = item_dict['code']
-        try:
-            item = Item.query.filter_by(code=code).one()
+        item = Item.query.filter_by(code=code).first()
+        if item is not None:
             item.update(item_dict)
             db.session.commit()
-        except NoResultFound as nrf:
-            pass
         return item
 
-    def delete_item(self, item_id):
-        try:
-            item = Item.query.filter_by(item_id=item_id).one()
-            db.session.delete(item)
-            db.session.commit()
-        except NoResultFound as nrf:
-            raise NoResultFound
+    def delete_item(self, item):
+        item = Item.query.filter_by(code=item.code).first()
+        if item is not None:
+            try:
+                db.session.delete(item)
+                db.session.commit()
+            except:
+                db.session.rollback()
+                raise
+            return True
